chore: setup release structure
