+++ packages/cli/package.json
@@ -10,24 +10,29 @@
   ],
   "type": "module",
   "scripts": {
-    "build": "node scripts/generate-version.js && tsc && pnpm run build:exe",
+    "build": "node scripts/generate-version.js && tsc",
     "build:clean": "sh ../../scripts/clean.sh && rm -rf binaries && pnpm run build",
     "build:release": "pnpm run build:clean",
-    "build:exe": "sh scripts/build-exe.sh all",
-    "build:exe:darwin-x64": "sh scripts/build-exe.sh darwin-x64",
-    "build:exe:darwin-arm64": "sh scripts/build-exe.sh darwin-arm64",
-    "build:exe:linux-x64": "sh scripts/build-exe.sh linux-x64",
-    "build:exe:linux-arm64": "sh scripts/build-exe.sh linux-arm64",
-    "build:exe:windows-x64": "sh scripts/build-exe.sh windows-x64",
+    "build:bin": "tsc && sh scripts/build-exe.sh all",
+    "build:bin:clean": "sh ../../scripts/clean.sh && rm -rf binaries && pnpm run build:bin",
+    "build:bin:release": "pnpm run build:bin:clean && pnpm run build:bin",
+    "build:bin:darwin-x64": "sh scripts/build-exe.sh darwin-x64",
+    "build:bin:darwin-arm64": "sh scripts/build-exe.sh darwin-arm64",
+    "build:bin:linux-x64": "sh scripts/build-exe.sh linux-x64",
+    "build:bin:linux-arm64": "sh scripts/build-exe.sh linux-arm64",
+    "build:bin:windows-x64": "sh scripts/build-exe.sh windows-x64",
     "lint": "eslint \"src/**/*.{js,ts}\" \"./**/__tests__/**/*.{js,ts}\"",
     "lint:fix": "eslint \"src/**/*.{js,ts}\" \"./**/__tests__/**/*.{js,ts}\" --fix",
     "test": "vitest run --config=./vitest.config.ts",
     "test:watch": "vitest --config=./vitest.config.ts",
-    "release": "pnpm run build:clean && pnpm publish",
+    "release": "pnpm run release:normal && pnpm run release:bin",
+    "release:normal": "pnpm run build:clean && pnpm publish",
+    "release:bin": "pnpm run prep-bin-release && pnpm run build:bin:clean && pnpm publish --tag bin && pnpm run restore-bin-release && pnpm run build:clean",
     "release:alpha": "pnpm run build:clean && pnpm publish --tag alpha",
     "release:beta": "pnpm run build:clean && pnpm publish --tag beta",
-    "release:bin": "pnpm run build:clean && pnpm publish --tag bin",
-    "release:latest": "pnpm run build:clean && pnpm publish --tag latest"
+    "release:latest": "pnpm run build:clean && pnpm publish --tag latest",
+    "prep-bin-release": "node scripts/prepare-binary-release.js",
+    "restore-bin-release": "node scripts/restore-package-json.js"
   },
   "exports": {
     ".": {
